PMD 7 upgrade
- now there is a good solution available to use PMD 7 with the Maven plugin (`pmd-compat6`)

+++ game/src/main/java/com/scheible/risingempire/game/api/view/colony/AnnexationStatusView.java
@@ -26,7 +26,7 @@ public record AnnexationStatusView(Optional<Integer> siegeRounds, Optional<Integ
 	public Optional<Integer> getProgress() {
 		if (this.siegeRounds.isPresent() && this.roundsUntilAnnexable.isPresent()) {
 			return Optional.of(Math.round(this.siegeRounds().get()
-					/ (float) (this.siegeRounds().get() + this.roundsUntilAnnexable().get()) * 100.0f));
+					/ (this.siegeRounds().get() + this.roundsUntilAnnexable().get()) * 100.0f));
 		}
 		else {
 			return Optional.empty();
